requirements.txt
baidu-aip==2.2.0.0
certifi==2018.1.18
chardet==3.0.4
idna==2.6
requests==2.18.4
tornado==4.5.3
urllib3==1.22
unidecode